Move context propagation processor to `camel-tracing`
