+++ components/camel-observation/src/main/java/org/apache/camel/observation/MicrometerObservationSpanAdapter.java
@@ -147,6 +147,7 @@ public class MicrometerObservationSpanAdapter implements SpanAdapter {
         return DEFAULT_EVENT_NAME;
     }
 
+    @Override
     public void setCorrelationContextItem(String key, String value) {
         Baggage baggage = tracer.createBaggage(key);
         Span span = getTracingContext().getSpan();
@@ -156,6 +157,7 @@ public class MicrometerObservationSpanAdapter implements SpanAdapter {
         baggage.set(span.context(), value);
     }
 
+    @Override
     public String getContextPropagationItem(String key) {
         Span span = getTracingContext().getSpan();
         if (span == null) {
