+++ components/camel-tracing/src/main/java/org/apache/camel/tracing/SetCorrelationContextProcessor.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.camel.opentelemetry;
+package org.apache.camel.tracing;
 
 import org.apache.camel.AsyncCallback;
 import org.apache.camel.Exchange;
@@ -23,7 +23,6 @@ import org.apache.camel.Traceable;
 import org.apache.camel.spi.IdAware;
 import org.apache.camel.spi.RouteIdAware;
 import org.apache.camel.support.AsyncProcessorSupport;
-import org.apache.camel.tracing.ActiveSpanManager;
 import org.apache.camel.util.ObjectHelper;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -45,13 +44,13 @@ public class SetCorrelationContextProcessor extends AsyncProcessorSupport implem
     @Override
     public boolean process(Exchange exchange, AsyncCallback callback) {
         try {
-            OpenTelemetrySpanAdapter camelSpan = (OpenTelemetrySpanAdapter) ActiveSpanManager.getSpan(exchange);
+            SpanAdapter camelSpan = ActiveSpanManager.getSpan(exchange);
             if (camelSpan != null) {
                 String item = expression.evaluate(exchange, String.class);
                 camelSpan.setCorrelationContextItem(baggageName, item);
             } else {
                 // avoid spamming logs
-                LOG.debug("OpenTelemetry: Cannot find managed span for Exchange: {}", exchange);
+                LOG.debug("Cannot find managed span for exchange: {}", exchange);
             }
         } catch (Exception e) {
             exchange.setException(e);
