+++ parent/pom.xml
@@ -386,10 +386,7 @@
         <opensearch-version>2.11.0</opensearch-version>
         <opensearch-testcontainers-version>2.1.0</opensearch-testcontainers-version>
         <openstack4j-version>3.12</openstack4j-version>
-        <opentelemetry-version>1.41.0</opentelemetry-version>
-        <opentelemetry-log4j2-version>2.7.0-alpha</opentelemetry-log4j2-version>
-        <opentelemetry-incubator-version>1.41.0-alpha</opentelemetry-incubator-version>
-        <opentelemetry-alpha-version>1.23.1-alpha</opentelemetry-alpha-version>
+        <opentelemetry-instrumentation-bom-alpha-version>2.7.0-alpha</opentelemetry-instrumentation-bom-alpha-version>
         <optaplanner-version>9.44.0.Final</optaplanner-version>
         <paho-version>1.2.5</paho-version>
         <parson-version>1.1.7</parson-version>
